\name{EnsDbLite.Hsapiens.88}
\docType{package}

\alias{EnsDbLite.Hsapiens.88-package}
\alias{EnsDbLite.Hsapiens.88}
\alias{EnsDbLite.Hsapiens.88}


\title{Ensembl-based annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at Ensembl on
  Mon Apr 10 16:28:31 2017 and based on the GRCh38
}

\author{Nobody}



\examples{
## load the library
##library(EnsDbLite.Hsapiens.88)
## list the contents that are loaded into memory
ls('package:EnsDbLite.Hsapiens.88')
## show the db object that is loaded by calling it's name
EnsDbLite.Hsapiens.88

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
